alert_bot/requirements.txt
yfinance
pandas
numpy
streamlit
requests
google-cloud-storage
